+++ lib/dialog/mult_select_dialog.dart
@@ -122,6 +122,7 @@ class _MultiSelectDialogState<T> extends State<MultiSelectDialog<T>> {
     _selectedValues.addAll(widget.initialValue);
 
     for (int i = 0; i < _items.length; i++) {
+      _items[i].selected = false;
       if (_selectedValues.contains(_items[i].value)) {
         _items[i].selected = true;
       }